implemented defer for recommended products
