+++ project-base/storefront/components/Pages/Cart/CartContent.tsx
@@ -3,17 +3,13 @@ import { CartLoading } from './CartLoading';
 import { CartSummary } from './CartSummary';
 import { OrderAction } from 'components/Blocks/OrderAction/OrderAction';
 import { OrderSteps } from 'components/Blocks/OrderSteps/OrderSteps';
+import { DeferredRecommendedProducts } from 'components/Blocks/Product/DeferredRecommendedProducts';
 import { useDomainConfig } from 'components/providers/DomainConfigProvider';
 import { TIDs } from 'cypress/tids';
 import { TypeRecommendationType } from 'graphql/types';
 import useTranslation from 'next-translate/useTranslation';
-import dynamic from 'next/dynamic';
 import { useCurrentCart } from 'utils/cart/useCurrentCart';
 import { getInternationalizedStaticUrls } from 'utils/staticUrls/getInternationalizedStaticUrls';
-
-const RecommendedProducts = dynamic(() =>
-    import('components/Blocks/Product/RecommendedProducts').then((component) => component.RecommendedProducts),
-);
 
 export const CartContent: FC = () => {
     const { t } = useTranslation();
@@ -52,7 +48,7 @@ export const CartContent: FC = () => {
             />
 
             {isLuigisBoxActive && (
-                <RecommendedProducts
+                <DeferredRecommendedProducts
                     itemUuids={cart.items.map((item) => item.uuid)}
                     recommendationType={TypeRecommendationType.Basket}
                     render={(recommendedProductsContent) => (
