+++ project-base/storefront/utils/useDeferredRender.ts
@@ -68,6 +68,7 @@ const PRODUCT_PAGE_DEFER_ORDER = [
     'navigation',
     'mobile_menu',
     'comparison_and_wishlist_button',
+    'recommended_products',
     'accessories',
     'newsletter',
     'user_consent',
@@ -80,6 +81,7 @@ const CATEGORY_PAGE_DEFER_ORDER = [
     'product_list',
     'filter_panel',
     'sorting_bar',
+    'recommended_products',
     'last_visited',
     'autocomplete_search',
     'cart_in_header',
@@ -95,6 +97,7 @@ const HOMEPAGE_DEFER_ORDER = [
     'loaders',
     'blog_preview',
     'footer',
+    'recommended_products',
     'promoted_products',
     'last_visited',
     'autocomplete_search',
